requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

datasets==2.16.1
Flask==3.0.2
python-box==7.1.1
PyYAML==6.0.1
requests==2.31.0
streamlit==1.31.1
transformers==4.36.2
